+++ security/provenance/fs.c
@@ -1009,13 +1009,13 @@ static ssize_t prov_read_prov_type(struct file *filp, char __user *buf,
 
 	if (type_info->is_relation) {
 		if (type_info->id)
-			strlcpy(type_info->str, relation_str(type_info->id),
+			strscpy(type_info->str, relation_str(type_info->id),
 				PROV_TYPE_STR_MAX_LEN);
 		else
 			type_info->id = relation_id(type_info->str);
 	} else {
 		if (type_info->id)
-			strlcpy(type_info->str,
+			strscpy(type_info->str,
 				node_str(type_info->id),
 				PROV_TYPE_STR_MAX_LEN);
 		else